updated gitattributes to have .gif as binary
