+++ .gitattributes
@@ -4,3 +4,4 @@
 *.jpeg binary
 *.ico binary
 *.icns binary
+*.gif binary
