+++ app/assets/images/logos/quit-docker-windows.gif
