+++ app/assets/images/logos/restart-docker-windows.gif
