+++ app/libs/ToolbarInterface.js
@@ -136,6 +136,7 @@ class ToolbarInterface {
      * @calls {gigantum.start}
      */
     const checkIsDockerReadyCallback = response => {
+      console.trace('checkIsDockerReadyCallback');
       if (response.success) {
         gigantum.start(gigantumStartCallback);
       }
